refactor: components exports
