+++ auth-app/src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as LoginView } from './login';
+export { default as UserProfileView } from './profile';
