+++ auth-app/src/pages/profile/index.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 
-import UserProfileView from '../../components/profile';
+import { UserProfileView } from '../../components';
 
 const ProfilePage: FC = () => <UserProfileView />;
